Fix issues with zooming in/out on mobile
A number of issues were occurring with zooming and on mobile browsers, one of which was causing the 'keep shopping' button to appear. This fix 1) allows zooming and scaling to happen by the user, 2) stops long store names from making the page wider than the viewport, 3) does a better job of hiding the 'keep shopping' button. Font sizes are also slightly increased to prevent unwanted zooming on mobile when focusing on inputs